requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.4.0
scipy>=1.7.0
python-multipart>=0.0.6
python-dotenv==1.0.0
transformers>=4.30.0
torch>=2.0.0
tensorflow==2.14.0
openai==1.3.0 